requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
py==1.11.0
pytz==2021.3
retry==0.9.2
sqlparse==0.4.2
